+++ admin/inertia/pages/home.tsx
@@ -95,7 +95,7 @@ export default function Home(props: {
   const { data: easySetupVisited } = useSystemSetting({
     key: 'ui.hasVisitedEasySetup'
   })
-  const shouldHighlightEasySetup = easySetupVisited?.value !== 'true'
+  const shouldHighlightEasySetup = easySetupVisited?.value ? easySetupVisited?.value !== 'true' : false
 
   // Add installed services (non-dependency services only)
   props.system.services